Dev: a small refactor of SiteBaker to help pinpoint a deploy bug
