+++ baker/readme.md
@@ -1,0 +1,3 @@
+# Baker
+
+This folder contains the scripts for baking the site for deployment to Netlify.
