+++ db/graphersToGit/tasks.ts
@@ -4,7 +4,7 @@ import * as fs from "fs"
 import * as db from "db/db"
 import { Grapher, GrapherProgrammaticInterface } from "grapher/core/Grapher"
 import { isPresent, mapToObjectLiteral } from "grapher/utils/Util"
-import { getPublishedGraphersBySlug } from "site/server/bakeGraphersToImages"
+import { getPublishedGraphersBySlug } from "baker/GrapherImageBaker"
 import { LegacyGrapherInterface } from "grapher/core/GrapherInterface"
 import { CoreTable } from "coreTable/CoreTable"
 import parseArgs from "minimist"
